feat(ui): offer a granularity on an applied temporal group key
Completes #786's surface, contract and evidence.

The drawer's applied-key list carries a granularity control on any temporal key,
offering exactly the periods the route reports and rendering itself away on
every other column. Removing a key drops its granularity with it — one left
behind is refused by the server rather than ignored.

The `grouping` URL param gains a `gran` member beside `agg`, dropped when empty
so an untruncated grouping produces the link it always did. A period outside the
vocabulary, or one naming a column that is not a key, refuses the whole payload
the way every other member of that envelope does.

`groupingContract.test.ts` now covers the granularity, not only the key list —
both the vocabulary and `periods` on the capability object. The assertion is not
redundant despite the vocabulary being shared: it is what fails if either
package restates the union locally, which is the change that would let them
disagree.

The live-DB smoke suite reconciles a drilled month two ways, and asserts the
timezone pin **with its counterfactual** — the pinned form's instant is
identical under two session zones and the unpinned form's is not, which a
one-sided check would not have shown.

One pre-existing failure went with it: the grouped-read smoke test asserted a
raw `date` key returns rows, and this table's date columns hold one value per
calendar day — exactly the tree the cardinality guard refuses — so it returned
none and asserted nothing. At a month it returns groups and the value is still a
Date, which is what that test is about.

+++ packages/server/src/db/group-query-builder/build-column-capabilities-query.util.ts
@@ -52,7 +52,10 @@ const PERIOD_CAPABLE_TYPE_SQL_NAMES = [
  * is what makes a **derived** group key measurable: `pg_stats` describes the raw
  * column, so nothing in the catalogue counts the distinct months in it — but a
  * range does bound them, and the first and last histogram bound are that range
- * (#786). Two properties keep it safe. The `CASE` guards the cast: Postgres does
+ * (#786). The `::float8` is not decoration: `extract(epoch …)` is `numeric`, which
+ * `pg` hands back as a **string**, and a string would coerce its way through the
+ * arithmetic that reads it while comparing wrong. Two further properties keep it
+ * safe. The `CASE` guards the cast: Postgres does
  * not evaluate an unselected branch, so `histogram_bounds::text::text[]` — the
  * only way to read an `anyarray` — is never applied to a type it would fail on.
  * And the scalar subquery over an absent histogram yields `NULL`, which is the
@@ -100,7 +103,7 @@ export const buildColumnCapabilitiesQuery = ({
        coalesce(s.n_distinct, 0) AS "nDistinct",
        c.reltuples AS "relTuples",
        CASE WHEN bt.typname = ANY($5::text[])
-            THEN (SELECT extract(epoch FROM (max(b::timestamptz) - min(b::timestamptz))) / 86400
+            THEN (SELECT (extract(epoch FROM (max(b::timestamptz) - min(b::timestamptz))) / 86400)::float8
                     FROM unnest(s.histogram_bounds::text::text[]) AS b)
        END AS "spanDays"
   FROM pg_attribute a
